Renaming irc3.async to irc3.asynchronous.
